フロントエンドのディレクトリ構造を更新し、機能別ディレクトリを括弧で囲みました。また、ホームページと商品ページのコンポーネントを新規作成し、グローバルスタイルとAPIサービスを追加しました。不要なファイルを削除し、商品データの取得ロジックを整理しました。
