--- frontend-customer/app/layout.tsx
@@ -1,39 +1,0 @@
-import MainLayout from "@/components/layout/MainLayout";
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
-
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Observability Shop | 高品質な商品をお求めやすい価格で",
-  description:
-    "オブザーバビリティショップでは、高品質な電子機器や最新ガジェットを、お求めやすい価格で提供しています。",
-  icons: {
-    icon: [{ url: "/shop-favicon.svg", type: "image/svg+xml" }],
-  },
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        <MainLayout>{children}</MainLayout>
-      </body>
-    </html>
-  );
-}
