+++ frontend-customer/src/app/favicon.ico
