+++ frontend-customer/src/app/globals.css
