+++ frontend-customer/src/app/page.tsx
@@ -1,0 +1,6 @@
+// /app/page.tsx をホームページにリダイレクト
+import HomePage from "./(features)/home/page";
+
+export default function Home() {
+  return <HomePage />;
+}
